refactor: created AboutImg, TextAbout and SectionTitle components
this components were from the About components.
The AboutImg was the image of the about section and, the TextAbout was the text of the about section and the SectionTitle was the title, the num and line used to separate the hero from the about section, this component now can be reused for other section.